Replace initial load of user's channel members for all teams with a lazy load
